Update FlightSQL metadata locations, names and docs
