+++ arrow-flight/src/sql/mod.rs
@@ -27,7 +27,7 @@
 //! 2. Helpers for encoding and decoding FlightSQL messages: [`Any`] and [`Command`]
 //! 3. A [`FlightSqlServiceClient`] for interacting with FlightSQL servers.
 //! 4. A [`FlightSqlService`] to help building FlightSQL servers from [`FlightService`].
-//! 5. Structures to build responses for FlightSQL metadata APIs: [`SqlInfoList`]
+//! 5. Helpers to build responses for FlightSQL metadata APIs: [`metadata`]
 //!
 //! [Flight SQL]: https://arrow.apache.org/docs/format/FlightSql.html
 //! [Apache Arrow]: https://arrow.apache.org
@@ -37,6 +37,7 @@
 //! [`do_get`]: crate::flight_service_server::FlightService::do_get
 //! [`FlightSqlServiceClient`]: client::FlightSqlServiceClient
 //! [`FlightSqlService`]: server::FlightSqlService
+//! [`metadata`]: crate::sql::metadata
 use arrow_schema::ArrowError;
 use bytes::Bytes;
 use paste::paste;
@@ -98,12 +99,9 @@ pub use gen::UpdateDeleteRules;
 pub use gen::XdbcDataType;
 pub use gen::XdbcDatetimeSubcode;
 
-pub use sql_info::SqlInfoList;
-
-pub mod catalogs;
 pub mod client;
+pub mod metadata;
 pub mod server;
-pub mod sql_info;
 
 /// ProstMessageExt are useful utility methods for prost::Message types
 pub trait ProstMessageExt: prost::Message + Default {
